refactor(facade): remove dependency to rtts_assert
