+++ modules/angular2/test/core/compiler/element_injector_spec.js
@@ -1,5 +1,5 @@
-import {describe, ddescribe, it, iit, xit, xdescribe, expect, beforeEach, SpyObject} from 'angular2/test_lib';
-import {isBlank, isPresent, FIELD, IMPLEMENTS, proxy} from 'angular2/src/facade/lang';
+import {describe, ddescribe, it, iit, xit, xdescribe, expect, beforeEach, SpyObject, proxy} from 'angular2/test_lib';
+import {isBlank, isPresent, FIELD, IMPLEMENTS} from 'angular2/src/facade/lang';
 import {ListWrapper, MapWrapper, List} from 'angular2/src/facade/collection';
 import {ProtoElementInjector, PreBuiltObjects, DirectiveBinding} from 'angular2/src/core/compiler/element_injector';
 import {Parent, Ancestor} from 'angular2/src/core/annotations/visibility';
