+++ modules/angular2/test/test_lib/test_lib_spec.js
@@ -1,7 +1,7 @@
-import {describe, it, iit, ddescribe, expect, tick, async, SpyObject, beforeEach} from 'angular2/test_lib';
+import {describe, it, iit, ddescribe, expect, tick, async, SpyObject, beforeEach, proxy} from 'angular2/test_lib';
 import {MapWrapper, ListWrapper} from 'angular2/src/facade/collection';
 import {PromiseWrapper} from 'angular2/src/facade/async';
-import {IMPLEMENTS, proxy} from 'angular2/src/facade/lang';
+import {IMPLEMENTS} from 'angular2/src/facade/lang';
 
 class TestObj {
   prop;
